refactor(skills,artefact,scheduler,capture): extract test modules into sibling files
Reduces the four remaining >700-line hotspots by splitting their #[cfg(test)] mod tests blocks into dedicated sibling files, mounted via #[path]. Production code stays co-located so private-field access is unaffected. No behaviour changes.
