--- commit_msg.txt
@@ -1,12 +1,0 @@
-refactor(shm): extract SharedBuffer tests into sibling module
-
-buffer.rs (720 lines) drops to 553 lines by moving the 168-line
-#[cfg(test)] mod tests block into a sibling buffer_tests.rs mounted
-via #[cfg(test)] #[path = "buffer_tests.rs"] mod tests;
-
-Production types (SharedBuffer, BufferDescriptor, gc_orphans) and all
-private helpers (RegionHeader, is_header_expired, gc_orphans_scan) stay
-co-located to retain access to SharedBuffer::inner / read_header and
-BufferDescriptor::from_buffer which relies on private struct fields.
-
-61 + 2 shm tests continue to pass unchanged.
